Rework for comments
