+++ standalone-metastore/metastore-server/src/main/java/org/apache/hadoop/hive/metastore/ObjectStore.java
@@ -648,13 +648,6 @@ public class ObjectStore implements RawStore, Configurable {
     }
   }
 
-  private void reInstateTransaction() {
-    int openTransactionCount = openTrasactionCalls;
-    rollbackTransaction();
-    openTransaction();
-    openTrasactionCalls = openTransactionCount;
-  }
-
   @Override
   public void createCatalog(Catalog cat) throws MetaException {
     LOG.debug("Creating catalog {}", cat);
@@ -2613,7 +2606,6 @@ public class ObjectStore implements RawStore, Configurable {
         tabGrants = this.listAllTableGrants(catName, dbName, tblName);
         tabColumnGrants = this.listTableAllColumnGrants(catName, dbName, tblName);
       }
-      List<Object> toPersist = new ArrayList<>();
       List<MPartition> mParts = new ArrayList<>();
       List<List<MPartitionPrivilege>> mPartPrivilegesList = new ArrayList<>();
       List<List<MPartitionColumnPrivilege>> mPartColPrivilegesList = new ArrayList<>();
@@ -2623,8 +2615,6 @@ public class ObjectStore implements RawStore, Configurable {
               + dbName + "." + tblName + ": " + part);
         }
         MPartition mpart = convertToMPart(part, table, true);
-
-        toPersist.add(mpart);
         mParts.add(mpart);
         int now = (int) (System.currentTimeMillis() / 1000);
         List<MPartitionPrivilege> mPartPrivileges = new ArrayList<>();
@@ -2633,7 +2623,6 @@ public class ObjectStore implements RawStore, Configurable {
             MPartitionPrivilege mPartPrivilege = new MPartitionPrivilege(tab.getPrincipalName(), tab.getPrincipalType(),
                 mpart, tab.getPrivilege(), now, tab.getGrantor(), tab.getGrantorType(), tab.getGrantOption(),
                 tab.getAuthorizer());
-            toPersist.add(mPartPrivilege);
             mPartPrivileges.add(mPartPrivilege);
           }
         }
@@ -2644,14 +2633,13 @@ public class ObjectStore implements RawStore, Configurable {
             MPartitionColumnPrivilege mPartColumnPrivilege = new MPartitionColumnPrivilege(col.getPrincipalName(),
                 col.getPrincipalType(), mpart, col.getColumnName(), col.getPrivilege(), now, col.getGrantor(),
                 col.getGrantorType(), col.getGrantOption(), col.getAuthorizer());
-            toPersist.add(mPartColumnPrivilege);
             mPartColumnPrivileges.add(mPartColumnPrivilege);
           }
         }
         mPartPrivilegesList.add(mPartPrivileges);
         mPartColPrivilegesList.add(mPartColumnPrivileges);
       }
-      if (CollectionUtils.isNotEmpty(toPersist)) {
+      if (CollectionUtils.isNotEmpty(mParts)) {
         GetHelper<Void> helper = new GetHelper<Void>(null, null, null, true,
             true) {
           @Override
@@ -2662,7 +2650,9 @@ public class ObjectStore implements RawStore, Configurable {
 
           @Override
           protected Void getJdoResult(GetHelper<Void> ctx) {
-            reInstateTransaction();
+            List<Object> toPersist = new ArrayList<>(mParts);
+            mPartPrivilegesList.forEach(toPersist::addAll);
+            mPartColPrivilegesList.forEach(toPersist::addAll);
             pm.makePersistentAll(toPersist);
             pm.flush();
             return null;
